introduce "unlock" subcommand
